tilepicker.c: use floats for tile coordinates
Use floats for coordinates so we can postulate "fractional" tiles.
This will be useful later when we want to draw the globe with a
guaranteed number of subdivisions at very low zoom levels, to keep
the spherical character when the world is just one tile across.

Floats should be fine because they can represent multiples of two
without loss of precision, and all we do is halving, doubling, and
dividing by powers of two. We use ldexpf() for losslessly
manipulating the exponent.

+++ coord3d.h
@@ -53,7 +53,7 @@ tilepoint_to_xyz_precalc (const double world_size, const double cx, const double
 }
 
 static inline void
-tilepoint_to_xyz (const int x, const int y, const double world_size, const double cx_lon, const double sin_cy_lat, const double cos_cy_lat, float p[3])
+tilepoint_to_xyz (const float x, const float y, const double world_size, const double cx_lon, const double sin_cy_lat, const double cos_cy_lat, float p[3])
 {
 	// Find the lat/lon to which the tilepoint corresponds:
 	double lon = world_x_to_lon(x, world_size);
