+++ tiledrawer.h
@@ -1,6 +1,6 @@
 #ifndef TILEDRAWER_H
 #define TILEDRAWER_H
 
-void tiledrawer (int tile_x, int tile_y, int tile_wd, int tile_ht, GLuint texture_id, const struct quadtree_req *req, float p[4][3]);
+void tiledrawer (float tile_x, float tile_y, float tile_wd, float tile_ht, GLuint texture_id, const struct quadtree_req *req, float p[4][3]);
 
 #endif	/* TILEDRAWER_H */
